sample_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic[email]
psycopg2-binary
